% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAT_functions.R
\name{check_lengths}
\alias{check_lengths}
\title{Check Lengths of Elements}
\usage{
check_lengths(...)
}
\arguments{
\item{...}{Elements to check. Each element should be a vector or list.}
}
\value{
A logical value. Returns \code{TRUE} if all elements have the same length, \code{FALSE} otherwise.
}
\description{
This function checks whether all provided elements have the same length.
}
\details{
The function takes a variable number of arguments and checks whether all provided elements (vectors or lists) have the same length.
The lengths of all elements are compared, and if they are equal, the function returns \code{TRUE}; otherwise, it returns \code{FALSE}.
}
\examples{
# Example usage
vec1 <- c(1, 2, 3)
vec2 <- c("a", "b", "c")
vec3 <- c(TRUE, FALSE, TRUE)

check_lengths(vec1, vec2, vec3) # Returns TRUE

vec4 <- c(1, 2)
check_lengths(vec1, vec2, vec4) # Returns FALSE

}
