% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAT_functions.R
\name{est_ability_mle}
\alias{est_ability_mle}
\title{Maximum Likelihood Ability Estimation Function}
\usage{
est_ability_mle(responses, as, bs, cs, optim_method = "CG")
}
\arguments{
\item{responses}{Numeric vector. The observed item responses (0 or 1).}

\item{as}{Numeric vector. The item discrimination parameters.}

\item{bs}{Numeric vector. The item difficulty parameters.}

\item{cs}{Numeric vector. The item guessing parameters.}

\item{optim_method}{Character. The optimization method to be used in the \code{optim} function. Default is \code{"CG"}.}
}
\value{
A list containing:
\describe{
  \item{ability_est}{The estimated ability level.}
  \item{ability_est_se}{The standard error of the ability estimate.}
}
}
\description{
This function estimates the most likely ability level of an individual given item parameters and observed responses by maximizing the log-likelihood function.
}
\details{
The function calculates the ability estimate by maximizing the log-likelihood function using the \code{optim} function with the specified optimization method.
The log-likelihood function is defined based on the probability of a correct response given the ability level and item parameters.
}
\examples{
# Example usage
responses <- c(1, 0, 1, 1, 0)
as <- c(1.2, 1.0, 0.8, 1.1, 0.9)
bs <- c(-1, 0, 1, 2, -0.5)
cs <- c(0.2, 0.25, 0.15, 0.3, 0.1)
result <- est_ability_mle(responses, as, bs, cs)
print(result)

}
\seealso{
\code{check_lengths} for checking the lengths of input vectors, \code{prob} for calculating response probabilities.
}
