% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAT_functions.R
\name{initial_item}
\alias{initial_item}
\title{Select the Initial Item for CAT}
\usage{
initial_item(item_bank_df, initial_ability = 0)
}
\arguments{
\item{item_bank_df}{Data frame. The item bank containing the parameters of the items. It should have columns: `item_id`, `a`, `b`, `c`.}

\item{initial_ability}{Numeric. The initial ability estimate. Default is 0.}
}
\value{
A data frame containing the selected initial item and additional information for the test event.
}
\description{
This function selects the initial item for a Computerized Adaptive Test (CAT) based on the initial ability estimate and weighted distance of item b parameters.
}
\details{
The function calculates weights for each item based on the distance from the initial ability estimate. Items within 0.5 logits of the initial ability estimate receive weights inversely proportional to their distance from the initial ability. The weights are normalized to sum to 1, and an item is randomly selected based on these weights. The returned data frame includes placeholders for response data and the current ability estimate.
}
\examples{
# Example usage
item_bank_df <- data.frame(
  item_id = 1:10,
  a = rlnorm(10, meanlog = 0.1, sdlog = 0.2),
  b = rnorm(10, mean = 0, sd = 1.5),
  c = runif(10, min = 0, max = 0.35)
)
result <- initial_item(item_bank_df, initial_ability = 0)
print(result)

}
