% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAT_functions.R
\name{item_info}
\alias{item_info}
\title{Item Information Function}
\usage{
item_info(theta, a, b, c)
}
\arguments{
\item{theta}{Numeric. The ability level of the examinee.}

\item{a}{Numeric. The discrimination parameter of the item.}

\item{b}{Numeric. The difficulty parameter of the item.}

\item{c}{Numeric. The guessing parameter of the item.}
}
\value{
A numeric value representing the information provided by the item at the given ability level (\eqn{\theta}).
}
\description{
This function calculates the item information for a given ability level (\eqn{\theta}) 
based on the parameters of an item response theory (IRT) model.
}
\details{
The item information function is based on the probability of a correct response given by the IRT model. 
The probability function (\code{prob}) is assumed to be defined elsewhere in the code. 
The information is calculated using the following formula:
\deqn{I(\theta) = a^2 \frac{Q}{P} \left( \frac{P - c}{1 - c} \right)^2}
where \eqn{P} is the probability of a correct response, \eqn{Q} is \eqn{1 - P}, \eqn{a} is the discrimination parameter, 
\eqn{b} is the difficulty parameter, and \eqn{c} is the guessing parameter.
}
\examples{
# Example usage
theta <- 1.5
a <- 1.2
b <- 0.5
c <- 0.2
item_info(theta, a, b, c)

}
\seealso{
\code{prob} for the probability function used in the calculation.
}
