% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAT_functions.R
\name{item_info_bank}
\alias{item_info_bank}
\title{Item Information for Item Bank}
\usage{
item_info_bank(item_bank, range = c(-3, 3), n_thetas = 100)
}
\arguments{
\item{item_bank}{Data frame. The item bank containing the parameters of the items. It should have columns named `a`, `b`, and `c` for the item parameters.}

\item{range}{Numeric vector of length 2. The range of ability levels (\eqn{\theta}) over which to calculate the item information. Default is \code{c(-3, 3)}.}

\item{n_thetas}{Integer. The number of ability levels (\eqn{\theta}) to evaluate within the specified range. Default is 100.}
}
\value{
A data frame containing the ability levels (\eqn{\theta}), item identifiers, and the corresponding item information values.
}
\description{
This function calculates the item information for each item in an item bank across a specified range of ability levels (\eqn{\theta}).
}
\details{
The function calculates the item information for each item in the item bank across a sequence of ability levels specified by the `range` and `n_thetas` parameters.
The item information is calculated using the \code{item_info} function.
}
\examples{
# Example usage
item_bank <- data.frame(
  a = rlnorm(10, meanlog = 0.1, sdlog = 0.2),
  b = rnorm(10, mean = 0, sd = 1.5),
  c = runif(10, min = 0, max = 0.35)
)
result <- item_info_item_bank(item_bank)
print(result)

}
\seealso{
\code{item_info} for the function used to calculate item information.
}
