% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAT_functions.R
\name{next_item}
\alias{next_item}
\title{Select the Next Item for CAT}
\usage{
next_item(eligible_items_df, test_event_df)
}
\arguments{
\item{eligible_items_df}{Data frame. The item bank containing the parameters of the eligible items. It should have columns: `item_id`, `a`, `b`, `c`.}

\item{test_event_df}{Data frame. The test event table containing the selected items and their corresponding information.}
}
\value{
A data frame containing the updated test event table with the next selected item.
}
\description{
This function selects the next item for a Computerized Adaptive Test (CAT) based on the current ability estimate.
}
\details{
The function selects the next item based on the distance between the current ability estimate and the item difficulty parameters (`b`). The item with the smallest distance is selected. The returned data frame includes placeholders for response data.
}
\examples{
# Example usage
eligible_items_df <- data.frame(
  item_id = 1:10,
  a = rlnorm(10, meanlog = 0.1, sdlog = 0.2),
  b = rnorm(10, mean = 0, sd = 1.5),
  c = runif(10, min = 0, max = 0.35)
)
test_event_df <- data.frame(
  order = integer(),
  item_id = integer(),
  a = numeric(),
  b = numeric(),
  c = numeric(),
  response_score = numeric(),
  current_ability = numeric(),
  current_ability_se = numeric(),
  item_selection_ts = as.POSIXct(character()),
  response = numeric(),
  response_ts = as.POSIXct(character())
)
test_event_df$current_ability <- 0
result <- next_item(eligible_items_df, test_event_df)
print(result)

}
