% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAT_functions.R
\name{prob}
\alias{prob}
\title{Probability Function}
\usage{
prob(theta, a, b, c)
}
\arguments{
\item{theta}{Numeric. The ability level of the examinee.}

\item{a}{Numeric. The discrimination parameter of the item.}

\item{b}{Numeric. The difficulty parameter of the item.}

\item{c}{Numeric. The guessing parameter of the item.}
}
\value{
A numeric value representing the probability of a correct response.
}
\description{
This function calculates the probability of a correct response in an IRT model.
}
