% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAT_functions.R
\name{simulate_cat}
\alias{simulate_cat}
\title{Simulate a Computerized Adaptive Test (CAT)}
\usage{
simulate_cat(
  item_bank,
  abilities,
  seed = 123,
  max_items = 20,
  min_se = 0.5,
  min_items = 0,
  response_consistency = 1,
  mle_kludge = TRUE,
  silent = FALSE,
  optim_method = "CG"
)
}
\arguments{
\item{item_bank}{Data frame. The item bank containing the parameters of the items. It should have columns: `item_id`, `a`, `b`, `c`.}

\item{abilities}{Numeric vector. The true ability levels of the test takers.}

\item{seed}{Integer. The random seed for reproducibility. Default is 123.}

\item{max_items}{Integer. The maximum number of items to be administered. Default is 20.}

\item{min_se}{Numeric. The minimum standard error stopping criterion. Default is 0.5.}

\item{min_items}{Integer. The minimum number of items to be administered before considering other stopping criteria. Default is NULL.}

\item{response_consistency}{Numeric. A multiplier for the item discrimination parameter to adjust response consistency. Default is 1.}

\item{mle_kludge}{Logical. Whether to apply a kludge to the MLE estimation to prevent extreme ability estimates. Default is TRUE.}

\item{silent}{Logical. Whether to print the case # after completing each simulated CAT. Default is FALSE.}

\item{optim_method}{Character. The optimization method to be used in the \code{optim} function. Default is \code{"CG"}.}
}
\value{
A list containing:
\describe{
  \item{summary}{A data frame summarizing the results for each test taker.}
  \item{event}{A data frame containing the detailed test event information for all test takers.}
}
}
\description{
This function simulates a Computerized Adaptive Test (CAT) for a given set of abilities using the specified item bank and stopping criteria.
}
\details{
The function simulates the CAT process by selecting items based on the current ability estimate, scoring responses, and updating the ability estimate until the stopping criteria are met. The stopping criteria include a maximum number of items, a minimum standard error, and an optional minimum number of items.
}
\examples{
# Example usage
item_bank <- data.frame(
  item_id = 1:10,
  a = rlnorm(10, meanlog = 0.1, sdlog = 0.2),
  b = rnorm(10, mean = 0, sd = 1.5),
  c = runif(10, min = 0, max = 0.35)
)
abilities <- rnorm(5, mean = 0, sd = 1)
result <- simulate_cat(item_bank, abilities)
print(result)

}
