% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAT_functions.R
\name{stop_test}
\alias{stop_test}
\title{Determine Whether to Stop the CAT}
\usage{
stop_test(test_event_df, max_items, min_se, min_items = 0)
}
\arguments{
\item{test_event_df}{Data frame. The test event table containing the selected items and their corresponding information.}

\item{max_items}{Integer. The maximum number of items to be administered.}

\item{min_se}{Numeric. The minimum standard error stopping criterion.}

\item{min_items}{Integer. The minimum number of items to be administered before considering other stopping criteria. Default is NULL.}
}
\value{
A logical value indicating whether the CAT should be stopped (\code{TRUE}) or not (\code{FALSE}).
}
\description{
This function determines whether to stop the Computerized Adaptive Test (CAT) based on the maximum number of items, minimum standard error (SE), and optional minimum number of items.
}
\details{
The function checks whether the test should be stopped based on the specified stopping criteria. If the current standard error is less than or equal to the minimum SE and the number of administered items is greater than or equal to the minimum items, or if the maximum number of items has been reached, the test will be stopped.
}
\examples{
# Example usage
test_event_df <- data.frame(
  order = integer(),
  item_id = integer(),
  a = numeric(),
  b = numeric(),
  c = numeric(),
  response_score = numeric(),
  current_ability = numeric(),
  current_ability_se = numeric(),
  item_selection_ts = as.POSIXct(character()),
  response = numeric(),
  response_ts = as.POSIXct(character())
)
test_event_df <- rbind(test_event_df, data.frame(
  order = 1,
  item_id = 1,
  a = 1.2,
  b = -0.5,
  c = 0.2,
  response_score = 1,
  current_ability = 0,
  current_ability_se = 0.3,
  item_selection_ts = Sys.time(),
  response = 1,
  response_ts = Sys.time()
))
result <- stop_test(test_event_df, max_items = 10, min_se = 0.3, min_items = 1)
print(result)

}
