% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAT_functions.R
\name{update_eligible_items}
\alias{update_eligible_items}
\title{Update List of Eligible Items for CAT}
\usage{
update_eligible_items(eligible_items_df, test_event_df)
}
\arguments{
\item{eligible_items_df}{Data frame. An item bank containing the parameters of all eligible items. It should have a column `item_id`.}

\item{test_event_df}{Data frame. The test event table containing the selected items and their corresponding information. It should have a column `item_id`.}
}
\value{
A data frame containing the updated list of eligible items.
}
\description{
This function updates the list of eligible items for a Computerized Adaptive Test (CAT) by removing items that have already been administered.
}
\details{
The function filters out items from the item bank that have already been administered, based on the `item_id` column.
}
\examples{
# Example usage
item_bank_df <- data.frame(
  item_id = 1:10,
  a = rlnorm(10, meanlog = 0.1, sdlog = 0.2),
  b = rnorm(10, mean = 0, sd = 1.5),
  c = runif(10, min = 0, max = 0.35)
)
test_event_df <- data.frame(
  item_id = c(1, 3, 5),
  order = c(1, 2, 3),
  a = c(1.2, 0.8, 1.1),
  b = c(-1, 0, 1),
  c = c(0.2, 0.25, 0.15),
  response_score = c(1, 0, 1),
  current_ability = c(NA, NA, NA),
  current_ability_se = c(NA, NA, NA),
  item_selection_ts = Sys.time(),
  response = c(1, 0, 1),
  response_ts = Sys.time()
)
result <- update_eligible_items(item_bank_df, test_event_df)
print(result)

}
